move balance types
