+++ zk.js/src/balance/balance.ts
@@ -1,4 +1,3 @@
-import { BN } from "@coral-xyz/anchor";
 import { Utxo } from "../utxo";
 import { PublicKey } from "@solana/web3.js";
 import {
@@ -9,41 +8,8 @@ import {
 } from "../constants";
 import { Provider } from "../wallet";
 import { Poseidon } from "../types/poseidon";
-
-// TODO: add history (spentutxos)
-
-/**
- * We keep spent UTXOs in a separate type,
- * because we need to keep Balance up2date at
- * any time, and syncing spent UTXOs is expensive.
- */
-export type Balance = {
-  // key is token
-  // includes only unspent UTXOs
-  tokenBalances: Map<string, TokenBalance>;
-  // TODO: add programBalances
-};
-
-export type TokenBalance = {
-  amount: BN;
-  lamports: BN; // rent
-  data: TokenData;
-  utxos: Utxo[];
-};
-
-// from ctx
-export type TokenData = {
-  symbol: string;
-  decimals: BN;
-  isNft: boolean;
-  isNative: boolean;
-  mint: PublicKey;
-};
-
-export type SerializedTokenBalance = {
-  mint: string;
-  utxos: { utxo: string; index?: number }[];
-};
+import { TokenData, SerializedTokenBalance } from "../types";
+import { Balance, TokenBalance } from "../types/balance";
 
 export const isSPLUtxo = (utxo: Utxo): boolean => {
   return !utxo.amounts[UTXO_ASSET_SPL_INDEX].eqn(0);
