+++ zk.js/src/balance/index.ts
@@ -1,7 +1,3 @@
-export { type TokenBalance as TokenBalance_new } from "./balance";
-export { type Balance as Balance_new } from "./balance";
-export { type TokenData as TokenData_new } from "./balance";
-export { type SerializedTokenBalance } from "./balance";
 export {
   getTokenDataByMint,
   initTokenBalance,
