+++ zk.js/src/types/index.ts
@@ -4,3 +4,8 @@ export * from "./accounts";
 export * from "./data";
 export * from "./programParameters";
 export * from "./result";
+
+export { type TokenBalance as TokenBalance_new } from "./balance";
+export { type Balance as Balance_new } from "./balance";
+export { type TokenData as TokenData_new } from "./balance";
+export { type SerializedTokenBalance } from "./balance";
